Move validate command to a separate package
