+++ internal/commands/run.go
@@ -3,21 +3,18 @@
 package commands
 
 import (
-	"context"
 	"errors"
 	"fmt"
+	"github.com/cirruslabs/cirrus-cli/internal/commands/helpers"
 	"github.com/cirruslabs/cirrus-cli/internal/commands/logs"
 	"github.com/cirruslabs/cirrus-cli/internal/executor"
 	eenvironment "github.com/cirruslabs/cirrus-cli/internal/executor/environment"
 	"github.com/cirruslabs/cirrus-cli/internal/executor/instance/containerbackend"
 	"github.com/cirruslabs/cirrus-cli/internal/executor/options"
 	"github.com/cirruslabs/cirrus-cli/internal/executor/taskfilter"
-	"github.com/cirruslabs/cirrus-cli/pkg/larker"
-	"github.com/cirruslabs/cirrus-cli/pkg/larker/fs/local"
 	"github.com/cirruslabs/cirrus-cli/pkg/parser"
 	"github.com/cirruslabs/cirrus-cli/pkg/rpcparser"
 	"github.com/spf13/cobra"
-	"io/ioutil"
 	"log"
 	"os"
 	"strings"
@@ -41,75 +38,6 @@ var debugNoCleanup bool
 // Experimental features flags.
 var experimentalParser bool
 
-// envArgsToMap parses and expands environment arguments like "A=B" (set operation)
-// and "A" (pass-through operation) into a map suitable for use across the codebase.
-func envArgsToMap(arguments []string) map[string]string {
-	result := make(map[string]string)
-
-	const (
-		keyPart = iota
-		valuePart
-		maxParts
-	)
-
-	for _, argument := range arguments {
-		parts := strings.SplitN(argument, "=", maxParts)
-
-		if len(parts) == maxParts {
-			// Set mode: simply assign the provided value to key
-			result[parts[keyPart]] = parts[valuePart]
-		} else {
-			// Pass-through mode: resolve the value for the provided key and assign it (if any)
-			resolvedValue, found := os.LookupEnv(parts[keyPart])
-			if !found {
-				break
-			}
-			result[parts[keyPart]] = resolvedValue
-		}
-	}
-
-	return result
-}
-
-func readYAMLConfig(path string) (string, error) {
-	yamlConfig, err := ioutil.ReadFile(path)
-	if err != nil {
-		if os.IsNotExist(err) {
-			return "", nil
-		}
-		return "", err
-	}
-
-	return string(yamlConfig), nil
-}
-
-func readStarlarkConfig(ctx context.Context, path string, env map[string]string) (string, error) {
-	starlarkSource, err := ioutil.ReadFile(path)
-	if err != nil {
-		return "", err
-	}
-
-	lrk := larker.New(larker.WithFileSystem(local.New(".")), larker.WithEnvironment(env))
-	return lrk.Main(ctx, string(starlarkSource))
-}
-
-func readCombinedConfig(ctx context.Context, env map[string]string) (string, error) {
-	yamlConfig, err := readYAMLConfig(".cirrus.yml")
-	if err != nil {
-		return "", err
-	}
-
-	starlarkConfig, err := readStarlarkConfig(ctx, ".cirrus.star", env)
-	if err != nil {
-		if os.IsNotExist(err) {
-			return yamlConfig, nil
-		}
-		return "", err
-	}
-
-	return yamlConfig + "\n" + starlarkConfig, nil
-}
-
 func run(cmd *cobra.Command, args []string) error {
 	// https://github.com/spf13/cobra/issues/340#issuecomment-374617413
 	cmd.SilenceUsage = true
@@ -125,10 +53,11 @@ func run(cmd *cobra.Command, args []string) error {
 		eenvironment.BuildID(),
 		eenvironment.ProjectSpecific(projectDir),
 	)
-	userSpecifiedEnvironment := envArgsToMap(environment)
+	userSpecifiedEnvironment := helpers.EnvArgsToMap(environment)
 
 	// Retrieve the combined YAML configuration
-	combinedYAML, err := readCombinedConfig(cmd.Context(), eenvironment.Merge(baseEnvironment, userSpecifiedEnvironment))
+	combinedYAML, err := helpers.ReadCombinedConfig(cmd.Context(),
+		eenvironment.Merge(baseEnvironment, userSpecifiedEnvironment))
 	if err != nil {
 		return err
 	}
